requirements.txt
numpy==2.1.2
opencv_python==4.5
Pillow==10.3.0
pytesseract==0.3.10
streamlit==1.39.0
opencv-contrib-python==4.10.0.84
img2table==1.3.0
torchvision==0.20.0
torch==2.5.0
python-docx==1.1.2
streamlit-authenticator>=0.2.2
bcrypt>=4.0.1
PyYAML>=6.0.1
streamlit-option-menu>=0.4.0
firebase-admin>=6.4.0